Migrate test suite to BDD style with pytest-describe
Restructure all tests around the public API surface using
describe_/when_/it_ nesting. Tests organized by API boundary
with DRY principle — each behavior tested once at the parent
API where it's consumed. Includes coverage gap tests for
construction validation, compile caching, ainvoke, astream,
and boundary conditions.
