+++ pyproject.toml
@@ -16,6 +16,7 @@ dependencies = [
 dev = [
     "pytest>=7.0",
     "pytest-asyncio>=0.21",
+    "pytest-describe>=2.2",
     "ruff>=0.9",
     "mypy>=1.14",
     "pre-commit>=4.0",
@@ -30,6 +31,7 @@ packages = ["src/langgraph_events"]
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
 testpaths = ["tests"]
+describe_prefixes = ["describe_", "when_"]
 
 [tool.ruff]
 line-length = 88
@@ -58,7 +60,7 @@ ignore = [
 ]
 
 [tool.ruff.lint.per-file-ignores]
-"tests/**" = ["S101", "PLR2004", "SIM117", "PLC0415"]
+"tests/**" = ["S101", "PLR2004", "SIM117", "PLC0415", "PLR0915", "N802"]
 "examples/**" = ["T20", "PLR2004", "S101", "S307", "PLC0415", "E501"]
 
 [tool.ruff.lint.isort]
